requirements.txt
asgiref==3.8.1
Django==5.1.2
djangorestframework==3.15.2
freezegun==1.5.1
iniconfig==2.0.0
packaging==24.1
pluggy==1.5.0
psycopg==3.2.3
psycopg-binary==3.2.3
pytest==8.3.3
pytest-django==4.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
